Re-organize stdlib into tools, runtime utils, builtins and prelude
Document runtime types.

ref #146
