+++ src/stdlib/index.ts
@@ -1,0 +1,3 @@
+export * from './runtime';
+export * from './builtin';
+export * from './prelude';
